refactoring and creating types.h to store all our struct
