+++ Makefile
@@ -15,8 +15,8 @@ ENV = $(addprefix $(ENV_DIR), $(ENV_FILES))
 
 BUILTINS_DIR = $(SRC_DIR)builtins/
 BUILTINS_FILES = 	pwd.c env.c echo.c unset.c cd.c exit.c\
-					export.c export_utils_0.c  export_utils_1.c\
-					utils.c builtins.c #cd.c echo.c env.c exit.c unset.c
+					export.c export_utils.c \
+					utils.c builtins.c
 					
 BUILTINS = $(addprefix $(BUILTINS_DIR), $(BUILTINS_FILES))
