+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: malde-ch <malo@chato.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/07 17:00:22 by malde-ch          #+#    #+#             */
-/*   Updated: 2025/02/21 04:18:02 by malde-ch         ###   ########.fr       */
+/*   Updated: 2025/02/21 05:46:23 by malde-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,46 +22,11 @@
 # include <readline/history.h>
 # include <limits.h>
 
-typedef struct s_env_var
-{
-	char				*key;
-	char				*value;
-	struct s_env_var	*next;
-}	t_env_var;
+# include "types.h"
 
-typedef enum e_operator
-{
-	OP_NONE, // 0
-	OP_AND, // &&
-	OP_OR, // ||
-	OP_PIPE, // |
-	OP_REDIRECT_OUT, // >
-	OP_REDIRECT_OUT_APPEND, // >>
-	OP_REDIRECT_IN, // <
-	OP_HERE_DOC, // <<
-}	t_operator;
-
-typedef struct s_redirection
-{
-	char		*file;
-	t_operator	type;
-}	t_redirection;
-
-typedef struct s_cmd
-{
-	char			**cmd;
-	t_redirection	*input;
-	t_redirection	*output;
-	struct s_cmd	*next;
-}	t_cmd;
-
-typedef struct s_mini
-{
-	int			exit_status;
-	char		**envp;
-	t_env_var	*env;
-	t_cmd		*cmd;
-}	t_mini;
+# include "exec.h"
+# include "builtins.h"
+# include "env.h"
 
 // to put in the parsing.h
 t_env_var	*parser_env(char **env);
@@ -75,9 +40,5 @@ void		free_all(t_mini *mini);
 
 // From main.c
 void		ft_exit(t_mini *mini, int num, char *str);
-
-# include "exec.h"
-# include "builtins.h"
-# include "env.h"
 
 #endif
